pyproject.toml
[project]
name = "tcrgnn"
version = "0.1.0"
description = "GNN TCR Python Package"
authors = [
    {name = "Amos Choo", email = "amos.choo@uq.edu.au"},
]
dependencies = [
    "torch_geometric==2.5.3",
    "jupyter",
    "matplotlib",
    "seaborn",
    "networkx",
    "biopython",
    "torch"
]
requires-python = ">=3.11"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"


[tool.pdm]
distribution = true

[tool.ruff]
line-length = 88
target-version = "py311"
exclude = [
    "old_folders",
]

# Enable Ruff formatting (replaces Black)
[tool.ruff.format]
quote-style = "double"       # or "single"
docstring-code-format = true
[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "F",   # pyflakes
    "W",   # warnings
    "I",   # import sorting
    "UP",  # pyupgrade
    "B",   # bugbear
    "C4",  # comprehensions
]
ignore = [
    "E501",  # optionally ignore line length if annoying
]
fixable = ["ALL"]




[dependency-groups]
dev = [
    "ruff>=0.14.2",
    "pre-commit>=4.3.0",
]
